fix(cli): patch security vulnerability in https-proxy-agent (npm advisory 1184) (#4603)
* chore(cli): vendor in all dependencies that depend on vulnerable package

Package `https-proxy-agent` has a security vulnerability which has
required a major version bump to address: https://hackerone.com/reports/541502

We depend on this package via 2 other packages (`proxy-agent` and
`pac-proxy-agent`), both of which have had a GitHub PR applied to bump
their dependencies on the vulnerable package, but both of which have
not released that change.

Vendor in both packages with updated dependencies to eliminate the
dependency on the vulnerable version of `https-proxy-agent`.

Internal reference: t/P29683837

* bundle vendored deps and omit vendor/ from tarball

* update package.json and packgae-lock.json