feat(ec2): Support explicit Subnet selection (#4622)
* feat(ec2): Support explicit Subnet selection

__WHY__
AWS CDK allows developers to attach subnets to other resource (e.g. ASG)
using Subnets selection feature. It implies re-use of existing subnets.
It is not possible to create subnet and attach it to ASG in same stack.

__WHAT__
The Subnet selection feature is expanded with ability to explicitly
define desired subnets. The subnet selection essentially becomes an
identity funnction.

* feat(ec2): incorporate code review comments