chore: require docstring on public members
Many people (still) contribute PRs with undocumented members, or
undocumented classes, or undocumented defaults, because there is no tool
support to check they're not forgetting.

We used to have `@default` checking on props types for L2s, but not for
interfaces that are used INSIDE L2 props structs, or that are used as
props types for non-constructs.

Add two new checks:

* Requirement to document top-level types and public members.
* Requirement to supply `@default` for optional properties on structs.

Which supersede the previous one.

Obviously, we have many currently undocumented APIs which are added
to the ignore list.

Fixes #2286.