fix(iam): fix `managedPolicyName`, cross-account references (#4630)
`ManagedPolicy#managedPolicyName` used to return the ARN instead of the
name. Since there is no way to get the name from CloudFormation, we use
ARN manipulation functions to get it.

The hard-named ARN of a ManagedPolicy used to contain the resource type
'role', due to a copy/paste mistake. Fix to use 'policy'.

Fixes #4581, fixes #4567.