fix(cloudformation): cannot reference resource attributes with "." in nested stacks (#4684)
When rendering the "GetAtt" to reference an auto-generated nested stack output
we did not use the actual `output.logicalId`, which can differ from the `id` passed
into the output (for example, periods are omitted since they are not supported in logical IDs).

This change uses the actual logical ID, but then requires that for sibling-references
we resolve this string since it's used for the logical ID of the parameter. Yeah, tricky.