feat(apigateway): async lambda proxy
Allows to invoke a Lambda function asynchronously with the raw request (proxy)
without having to fully setup a custom integration (templates).

This is useful for applications that require work to be performed asynchronously
but would like to receive the raw request easily.

Uses a singleton Lambda function which invokes the target Lambda function with
the `Event` invocation type.