chore: `stringifyNumber` handles non-numbers
In non-typechecked languages such as JavaScript and Python, `stringifyNumber` can easily be passed a non-number, for example a Token from a `CfnParameter`. The previous implementation would  straight-up stringify that object and the value would be lost.

Since it is easy to allow that case anyway, add support for customers that don't hold to the contract.

Fixes #4656