fix(init/java): correct generated unit test
The unit test from the "app" init template for Java would perform unsafe
operations on the result of an `any` typed method. This used to return
a `java.util.Map` (which was unsafe), and now returns an Object
reference, which can no longer be turned to JSON using a plain Jackson
serializer as it was doing before.