fix(ssm): malformed ARNs for parameters with physical names that use path notation (#4842)
* fix(ssm): malformed ARNs for parameters with physical names that use path notation

SSM parameter names can have one of two forms: “simpleName” or “/path/name”. This makes it tricky to render an ARN for the parameter is the name is an unresolvable token (such as a “Ref”) because we can’t decide whether a “/“ separator is required in the ARN. The previous implementation assumed "Ref" always returns the name without a "/" prefix, and therefore did not use the "/" separator. This fix will use the physical name itself (if possible) to determine the separator (and also assume that generated names will not use the path notation).

The only case where this is impossible is if the physical name is a token (either created or imported), in which case we should be able to synthesize a CloudFormation condition which will parse the token during deployment.

This test also adds a validation that verifies that if a physical name is provided and uses path notation, it must begin with a "/".

Misc: re-add `install.sh` to call `npx yarn install`

* explicit parameterArnSeparator

revert attempt to guess parameter name prefix if it's a token since we can't incorporate refs in conditions. Instead, if the parameter name
if a token, we expect `parameterArnSeparator` to be explicitly defined and be one of "/" or "".

* misc

* fix test expectation
* add public API doc

* add --frozen-lockfile to install.sh

* rename "parameterArnSeparator: string" to "simpleName: boolean"