user-defined handlers are now passed as lambda.IFunction and make isComplete optional
Instead of presuming to define the AWS Lambda function on behalf of the user, we now simply ask the user to supply a couple of IFunction objects. This gives users maximum flexibility in configuring their AWS Lambda environments.

Furthermore, this change also makes isComplete optional and only defines the waiter resources in case isComplete is provided.

Refactored integ test so that providers are deployed as a separate stack and their entrypoints are exported through CFN export names.