stop calling isComplete synchronously
Since we now have a better way to represent that onEvent is synchronous (by simply not supplying an isComplete handler), and this will also result in a cleaner stack (no step functions, etc), the value of calling isComplete synchronously has been reduced. Combined with the observation by @jogold that the framework onEvent timeout would have to be user onEvent+isComplete, the complexity is not worth it.

Also, refactor unit test harness such that it simulates the entire workflow (incl. reaction to startExecution and simulation of the waiter step function).