chore: "compat" script for api compatibility check
Add a `compat` script to all modules and to the repo root which runs api compatibility checks.
Also add a section in the contribution guide that explains about api checks.

Add a "build" and "pack" scripts in the root package.json so users can just run "yarn build/pack" from the root.