chore(build-tools): prefix file filter in cdk-test
Treat non-option arguments passed to cdk-test as prefix file filters.
When run without this argument, no filter is applied, i.e., all tests
are run.

This is useful for packages that have a large number of unit tests, such
as 'core'. When focused on a specific area of such a package, it allows
for running only specific tests that are of interest; thereby reducing
distraction, reducing scrollback and faster iteration.