feat(lambda-layers): add AWS SDK JS layer
Add a new package `@aws-cdk/lambda-layers` that offers pre-built layers.

Layer building is currently only supported for Node.js and works as follows:
* List static members of the `Layer` class (= layers to build)
* Create a dummy `package.json` file containing only dependencies for this layer
in the `layers` folder. Dependencies versions are extracted from the `package.json`
of `@aws-cdk/lambda-layers`.
* Run `npm install`

As layer dependencies are extracted from the main `package.json` they can be picked
up by `dependabot`, automatically updating the layer when needed.

The `getLayerVersion()` method returns a singleton `lambda.LayerVersion` that can be
used in Lambda functions.

The first layer is `AWS_SDK_JS` which offers a more recent version of the AWS SDK than
the one included in the Lambda runtime (it will also be updated more frequently).

Use the `AWS_SDK_JS` layer in `AwsCustomResource` and `AwsApi` event target.

Closes #2689
Closes #5063