docs(assert): stop advertising integ tests in `assert` (#5159)
The `@aws-cdk/assert` package was advertising being able to run
Integration Tests, and this has prompted people to try and use it
that way.

This was untrue, the integration test facilities are not in the
`assert` package but in the `cdk-integ-tools` package, which is
currently not reusable.

Remove the false advertising.