feat: Make activating Python virtualenv easier on Windows (#5014)
* Make activating virtualenv easier on Windows

The command for activating a Python virtualenv on Linux or Mac OS X is: `source .env/bin/activate`

The command for doing it on Windows is: `.env\Scripts\activate.bat`

This adds `source.bat` to our Python templates. It's a simple batch file that lets the Linux/Mac virtualenv activating command work on Windows. That way, when we tell users to activate their virtualenv in documentation, we need only give the Linux/Mac command, and it'll work even on Windows. (Windows `cmd.exe` doesn't have a `source` command, so this doesn't block any functionality that Windows users are accustomed to.)

* Add source.bat to sample-app template as well

* Add message about what the script is doing

* Add message indicating what the script is doing for you