fix(route53): return plain hosted zone id without /hostedzone/ prefix
When creating a HostedZone, the `hostedZoneId` is the ID of this hosted zone, such as "Z23ABC4XYZL05B".

Until how, when importing a HostedZone using route53.HostedZone.fromLookup, it did return the `hostedZoneId` in this format: /hostedzone/Z23ABC4XYZL05B.

This commit makes the value of `hostedZoneId` consistent between creating a new `HostedZone`, using `HostedZone.fromLookup` and `fromHostedZoneAttributes`.

BREAKING CHANGE: value of `hostedZoneId` is now be different compared to previous versions of CDK when using  `HostedZone.fromLookup` or `fromHostedZoneAttributes`