fix(core): dependencies across stack boundaries of all kinds (#5211)
* fix(cloudformation): dependencies across nested stack boundaries

When adding dependencies of any kind (`node.addDependency`, `stack.addDependency` or `cfnResource.addDependsOn`), we now take nested stacks into account and transfer the dependency to the lowest common scope (either a common stack or the assembly).

This is primarily implemented in `cfnResource.addDependsOn` and `stack.addDependency` through the following algorithm:

- Lookup the lowest common stack
- If not found, apply the dependency at the assembly level (between two top-level stacks).
- If found, identify the two CloudFormation resources which represent this relationship within the scope of the common stack (this could be the actual resource if they are both in the same stack or the `AWS::CloudFormation::Stack` resource if the source or target are with in a nested stack).
- Apply the dependency at the CloudFormation level (add `DependsOn`).

There are a bunch of edge cases as well.

To be able to find the AWS::CloudFormation::Stack resource for nested stacks, the `nestedStackResource` property was added to `core.Stack` and assigned by `NestedStack`. It seems like we should probably move `NestedStack` from the `aws-cloudformation` module and into `core` since there is too much coupling right now.

Created a thorough test suite to examine various cases under `test.deps.ts`. Note that this suite includes tests for nested stacks and for regular cases, but decided to put them all in a single suite to make it easier to reason about coverage.

Fixes #4460

* validate #4474

* combine stack and resource deps into a single algorithm

- Rewrote the `addDependency` algorithm generically to support defining dependencies between two resources or two stacks.
- Added a cache for `Stack.of` which is expected to have some nice gains since this method is widely used, especially around these algorithms.

* rename `parentStack` to `nestedStackParent`

* hide the Stack.of cache property

* convert nestedStackParent to a getter

* fix compilation error