fix(route53): return plain hosted zone id without /hostedzone/ prefix (#5230)
When creating a HostedZone, the `hostedZoneId` is the ID of this hosted zone, such as "Z23ABC4SAMPLE".

Until how, when importing a HostedZone using route53.HostedZone.fromLookup, it did return the `hostedZoneId` in this format: /hostedzone/Z23ABC4SAMPLE.

This commit makes the value of `hostedZoneId` consistent between creating a new `HostedZone`, using `HostedZone.fromLookup` and `fromHostedZoneAttributes`.

BREAKING CHANGE: the value of `hostedZoneId` will no longer include `/hostedzone/` prefix and only includes the hostedZoneId when using  `HostedZone.fromLookup` or `fromHostedZoneAttributes`