chore: normalize line endings of .bat files
The files were committed with CRLF line endings to the git object store
(see `git cat-file -p ae6356a7c426aad1d439d22221bba6f1480673c2 | xxd`).

Unsure how this happened, since .gitattributes should have prevented it.
Maybe it was an old version of git that relied on `core.autocrlf` only,
which defaults to `false`?

On machines that respect .gitattributes' `text=auto` setting, git now
wants to convert the literal CRLF in the working copy to LF for the
object store, and so the file shows up as different.

Normalize line endings in the .bat files to shop this from showing
up as changed.