fix(codebuild): rename CodeBuild VPC policy to be unique (#5385)
The policy used for the VPC permissions had the hard-coded name 'CodeBuildEc2Polocy',
which meant you could only have at most one CodeBuild project in a VPC in each stack.
The change is to use the construct's unique ID as the policy name,
which means it will be different for every project in the stack.