feat(lambda): asynchronous lambda chain
Add a `LambdaChain` construct to chain asynchronous invocations of Lambda
functions.

Asynchronously invoking the first function starts the execution chain. The
response payload of a successful invocation of a function in the chain is
passed as request payload to the next function automatically using
EventBridge.

Allows to start a series of long running processes.

The poor man's step function?