feat(assets): asset bundlers
Introduce bundlers for S3 assets and implement Parcel for JavaScript/TypeScript
entry files. Add a `fromParcel` static method to `@aws-cdk/aws-lambda.Code` to
create Parcel bundled Lambda functions.

Bundlers are defined in `@aws-cdk/cx-api` and extend the `Bundler` abstract
class. They implement a `bundle` method that takes a `FileAssetMetadataEntry`
and returns a transformed `FileAssetMetadataEntry`. The zip packaging is then
applied on the transformed asset.

Since assets are currently "transported" to `aws-cdk` using metadata, concrete
TypeScript class objects cannot be "transported". To overcome this limitation
the class name and constructor options of the bundler are used as metadata. A
new instance of the bundler is then "re-created" in `aws-cdk` where the real
bundling occurs. This has the disadvantage that bundlers can only be implemented
in `@aws-cdk/cx-api` (user provided bundlers will not work).