fix(eks): generated cluster name can exceed 63 characters
Since the generated cluster name includes the logical ID of the resource as a prefix with a postfix of the request ID, the resulting generated name can exceed 63 characters.

Fixes #5596

NOTE: since the current version of the EKS module have not been released yet, this is not a breaking change.