chore(eks): kubectl resource provider (#5625)
* chore(eks): kubectl resource provider

Merge the `Helm` and `KubernetesResource` providers into a single "KubectlProvider" which uses the resource type to decide which actual provider will be used. This is in order to reduce the footprint of this module and since we will need to introduce an additional provider in order to support Fargate Profiles.

* update unit test expectations

* update test expectations