chore(cli): test CDK can update an existing VPC
Add an integration test with a stack that doesn't get cleaned up,
but updated on every test run.

This makes sure that new releases of the CDK can still update
stacks that have been deployed using previous CDK releases without
issue. We only test shared, long-lived resources which would be
a pain to drop and recreate if a CDK upgrade would break deploying
to them, like a VPC.

Fixes #5682.