fix(ecr-assets): unable to use one Dockerfile to build multiple images (#5705)
* fix(ecr-assets): unable to use one Dockerfile to build multiple images

Since we use the source hash as the image ID, if we use different docker build options such as specifying a different `target`, `buildArgs` or a custom docker file name, we still get the same image ID. This means that if this image ID already exists, we skip the build. This makes it impossible to use the same docker build context to build multiple images.

This fix exposes the ability to include extra information to the source hash fingerprint algorithm (in `assets.Staging`), and utilizes this capability to "salt" the hash with build arguments if they are provided.

Fixes #5683

* code review feedback

- rename `extra` to `extraHash` so it makes more sense when extending `FingerprintOptions`.
- build `extraHash` as an object instead of an array.
- include `repositoryName` in fingerprint

* Update fingerprint.ts

* Update copy.ts

* fix build

* add test for repository name

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>