chore: fix repeated build of aws-events-targets (#5708)
Since we are `require()`ing the `sdk-api-metadata.json` file, TypeScript
is going to generate a type definition `.d.ts` file for it.

However, it forgets to add it to its own ignore list, so upon the
next compilation it will both try to read *and* generate '.d.ts' file,
leading it to think it's about to overwrite a source file with the
following error:

```
Cannot write file 'sdk-api-metadata.d.ts' because it would overwrite input file.
```

Delete the `.d.ts` file before each build so we don't run into
this build failure (until tsc fixes this issue upstream).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>